Parity with wiremod#2399
**DO NOT MERGE THIS UNTIL https://github.com/wiremod/wire/pull/2399 IS MERGED.**

This PR converts functions that return a table with three numbers to proper Vector/Angle userdata, since that is what vector and angle types will now use in E2 for efficiency.

I'm sweeping through found instances of this on github and making these PRs to make sure the amount of extensions broken by this is minimal.